+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.0.11
+- Fix for Issue #12
+- Add ON_LOADED event Issue #11
+
 ### 0.0.10
 - Updated FreKotlin
 - Updated Android Play Services to 11.4.2